refactor(craft): move sandbox celery tasks to background/celery/tasks/build
Repo convention puts celery tasks under background/celery/tasks/, not
inside feature packages. Also hoists the function-level imports to the
top of the module (no import cycle existed) and updates the heavy/light
worker autodiscover entries.
