+++ backend/onyx/background/celery/apps/heavy.py
@@ -131,8 +131,8 @@ celery_app.autodiscover_tasks(
     app_base.filter_task_modules(
         [
             "onyx.background.celery.tasks.pruning",
-            # Sandbox tasks (file sync, cleanup)
-            "onyx.server.features.build.sandbox.tasks",
+            # Sandbox tasks (file sync, cleanup; build feature)
+            "onyx.background.celery.tasks.build",
             "onyx.background.celery.tasks.hierarchyfetching",
         ]
     )
