+++ backend/onyx/background/celery/apps/light.py
@@ -159,8 +159,8 @@ celery_app.autodiscover_tasks(
             "onyx.background.celery.tasks.doc_permission_syncing",
             "onyx.background.celery.tasks.docprocessing",
             "onyx.background.celery.tasks.opensearch_migration",
-            # Sandbox cleanup tasks (isolated in build feature)
-            "onyx.server.features.build.sandbox.tasks",
+            # Sandbox cleanup tasks (build feature)
+            "onyx.background.celery.tasks.build",
         ]
     )
 )
