+++ docs/craft/legacy/sandbox-backends.md
@@ -27,7 +27,7 @@ V1 adds a third backend, `docker`, that the api_server drives directly via the D
 - **The api_server needs the Docker socket.** Mount `/var/run/docker.sock` into the api_server container in `docker-compose.yml`, behind a clearly-documented opt-in. Self-hosted admins who don't want to mount the socket can set `SANDBOX_BACKEND=local` and accept the lack of isolation. **Mounting the socket is equivalent to root on the host** — call this out explicitly in the docs and explain that this is the same trust model that Kubernetes' RBAC for sandbox pods provides.
 - **Networking.** One Docker bridge network per Craft deployment (e.g. `onyx_craft_sandbox`). The sandbox containers join it; the api_server joins it. NextJS dev servers are reachable by the api_server via the container's hostname (`sandbox-<id>:<port>`), the same way K8s reaches them via the cluster service DNS. We do **not** publish NextJS ports on the host — proxy through the api_server's existing webapp-proxy code path. This avoids host-port collisions across users, keeps the URL shape consistent with K8s (`http://<service>.<ns>.svc.cluster.local:<port>` ↔ `http://sandbox-<id>:<port>`), and keeps the sandbox off the host network.
 - **Snapshots go through `FileStore`, not s5cmd.** K8s uses an s5cmd-in-pod path because the api_server can't see the pod's filesystem and IRSA is convenient. In Docker we have direct access to the container. Use the existing `SnapshotManager` (already used by `local` infrastructure-side) — exec a `tar -czf -` inside the container, stream the bytes back to the api_server, hand them to `FileStore.save_file(...)` with `FileOrigin.SANDBOX_SNAPSHOT`. Restore reverses it. This works against MinIO, S3, GCS, or local file-store backends with zero extra config.
-- **Idle cleanup task.** `cleanup_idle_sandboxes_task` in `sandbox/tasks/tasks.py:50` currently checks `SANDBOX_BACKEND == LOCAL` and bails, then casts to `KubernetesSandboxManager`. Generalize this so docker mode also runs idle cleanup and snapshotting. The session-listing helper (`_list_session_directories` at `tasks.py:201`) needs a docker-aware path; ideally it moves onto the manager itself instead of being defined on K8s in the tasks module.
+- **Idle cleanup task.** `cleanup_idle_sandboxes_task` in `background/celery/tasks/build/tasks.py:50` currently checks `SANDBOX_BACKEND == LOCAL` and bails, then casts to `KubernetesSandboxManager`. Generalize this so docker mode also runs idle cleanup and snapshotting. The session-listing helper (`_list_session_directories` at `tasks.py:201`) needs a docker-aware path; ideally it moves onto the manager itself instead of being defined on K8s in the tasks module.
 - **`local` retained for dev, never marketed as secure.** Update `sandbox/README.md`, `configs.py`, and the env template to make the role of each backend explicit. Default for `docker compose` deployments shifts to `docker`. Default for plain `python -m onyx.main` (running outside containers) stays `local`.
 - **Backwards compat.** No migration. Per the main plan, V1 has no migration story for existing Craft sessions — they will be wiped on upgrade.
 - **Don't rename the K8s docker subtree.** The Dockerfile lives at `backend/onyx/server/features/build/sandbox/kubernetes/docker/Dockerfile`. That path is misleading once Docker (the SDK) is also a backend, but the main plan is explicit about not renaming `build/` modules in V1. Live with the awkward path; document it once in a comment.
@@ -169,7 +169,7 @@ api_server                              sandbox container
 
 **Existing files to touch:**
 - `backend/onyx/server/features/build/sandbox/base.py` — add `list_session_workspaces(sandbox_id) -> list[UUID]` to the ABC. Implement it on K8s and Local (the local version walks the directory; K8s's logic moves out of `tasks.py` onto the manager). The docker version execs `ls /workspace/sessions/`.
-- `backend/onyx/server/features/build/sandbox/tasks/tasks.py:67-101` — `cleanup_idle_sandboxes_task` no longer special-cases backends. Drop the `if SANDBOX_BACKEND == LOCAL: return` (replaced with: `manager.supports_idle_cleanup()` → bool, default False for local) and the `isinstance(KubernetesSandboxManager)` cast. `_list_session_directories` deleted; calls go through `manager.list_session_workspaces(...)`.
+- `backend/onyx/background/celery/tasks/build/tasks.py:67-101` — `cleanup_idle_sandboxes_task` no longer special-cases backends. Drop the `if SANDBOX_BACKEND == LOCAL: return` (replaced with: `manager.supports_idle_cleanup()` → bool, default False for local) and the `isinstance(KubernetesSandboxManager)` cast. `_list_session_directories` deleted; calls go through `manager.list_session_workspaces(...)`.
 - `backend/onyx/server/features/build/sandbox/snapshot_manager.py` — already takes a `FileStore` and works in tar/upload terms. Add a `persist_snapshot_from_stream(stream, ...)` and `restore_snapshot_to_stream(...)` so docker can hand it raw bytes from `exec_run` rather than a path on disk. Local uses the existing path-based methods unchanged.
 - `backend/onyx/server/features/build/sandbox/README.md` — section on `docker` backend, comparison table of the three backends, explicit warning that `local` is dev-only, and the trust-boundary note about the docker socket.
